Remove redundant version check and unit test case
This change introduces some small fixes for the `util/version` package of
`keel`. There should be no breaking changes introduced by the changes being
proposed.

The first change is around the version check when the policy is set to
PolicyTypePatch. In this path, the major and minor versions are confirmed to be
the same. It then checks to make sure the new patch is greater than the current
one, which we already do here:

* https://github.com/theckman/keel/blob/b3ba3a47f1c9e0f4fc93571645286aacf00b4d39/util/version/version.go#L129-L132

This change removes the check confirming that the patch is higher on the new
version. An additional test was added to test for a patch decrease, with
PolicyTypePatch, to make sure nothing was broken.

Lastly, in the process of writing this test a duplicated test case was
discovered in the test file so it was removed.

Signed-off-by: Tim Heckman <t@heckman.io>